Refactor sound player to bubble sound selection
