+++ app/assets/tailwind/themes/focus-space.css
@@ -80,43 +80,45 @@
     border-radius: 50%;
   }
 
-  .ambient-sounds-list {
+  .ambient-sounds-player {
     @apply max-w-3xl mx-auto;
+  }
 
-    .actions {
-      @apply flex grow mb-4 gap-8 text-xl justify-between;
+  .control-bar {
+    @apply flex grow mb-4 gap-8 text-xl justify-between;
 
-      .play-button {
-        @apply text-base flex items-center rounded-md px-4 py-2;
-        border: 1px solid color-mix(in srgb, var(--readable-color) 60%, transparent);
-        color: color-mix(in srgb, var(--readable-color) 60%, transparent);
+    .play-button {
+      @apply text-base flex items-center rounded-md px-4 py-2;
+      border: 1px solid color-mix(in srgb, var(--readable-color) 60%, transparent);
+      color: color-mix(in srgb, var(--readable-color) 60%, transparent);
 
-        svg {
-          @apply mr-2;
-        }
-
-        &:hover {
-          color: color-mix(in srgb, var(--readable-color) 80%, transparent);
-          border-color: color-mix(in srgb, var(--readable-color) 80%, transparent);
-        }
+      svg {
+        @apply mr-2;
       }
 
-      .preset-button {
-        @apply text-base;
-
-        color: color-mix(in srgb, var(--readable-color) 60%, transparent);
-
-        svg {
-          @apply mr-2;
-        }
-
-        &:hover {
-          color: color-mix(in srgb, var(--readable-color) 80%, transparent);
-        }
+      &:hover {
+        color: color-mix(in srgb, var(--readable-color) 80%, transparent);
+        border-color: color-mix(in srgb, var(--readable-color) 80%, transparent);
       }
-
     }
 
+    .preset-button {
+      @apply text-base;
+
+      color: color-mix(in srgb, var(--readable-color) 60%, transparent);
+
+      svg {
+        @apply mr-2;
+      }
+
+      &:hover {
+        color: color-mix(in srgb, var(--readable-color) 80%, transparent);
+      }
+    }
+
+  }
+
+  .ambient-sounds-list {
     .sound {
       @apply relative rounded-lg;
       border-width: 3px;
